# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for flight management system",
        "title": "Flight Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.ruaairline.com/support",
            "email": "support@ruaairline.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/params": {
            "get": {
                "description": "Retrieve a list of all parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parameters"
                ],
                "summary": "Get all the parameters",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Parameter"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update all parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parameters"
                ],
                "summary": "Update parameters",
                "parameters": [
                    {
                        "description": "Parameter information",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Parameter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Parameter"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/planes": {
            "get": {
                "description": "Retrieve a list of all planes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "planes"
                ],
                "summary": "Get all planes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PlaneResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/planes/{code}": {
            "get": {
                "description": "Retrieve a plane by its unique code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "planes"
                ],
                "summary": "Get plane by code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plane Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PlaneResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "dto.PlaneResponse": {
            "type": "object",
            "properties": {
                "plane_code": {
                    "type": "string"
                },
                "plane_name": {
                    "type": "string"
                }
            }
        },
...
        "models.Parameter": {
            "type": "object",
            "properties": {
                "latest_ticket_purchase_time": {
                    "type": "integer"
                },
                "max_intermediate_stop_duration": {
                    "type": "integer"
                },
                "max_intermediate_stops": {
                    "type": "integer"
                },
                "max_ticket_classes": {
                    "type": "integer"
                },
                "min_flight_duration": {
                    "type": "integer"
                },
                "min_intermediate_stop_duration": {
                    "type": "integer"
                },
                "number_of_airports": {
                    "type": "integer"
                },
                "ticket_cancellation_time": {
                    "type": "integer"
                }
            }
        }
    }
}